Use new tabs in Settings
[MAILPOET-2775]
