+++ assets/js/src/settings/index.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import 'parsleyjs';
 import ReactDOM from 'react-dom';
-import { HashRouter } from 'react-router-dom';
 import { GlobalContext, useGlobalContextValue } from 'context';
 import { initStore } from './store';
 import Settings from './settings';
 
 const Entry = () => (
   <GlobalContext.Provider value={useGlobalContextValue(window)}>
-    <HashRouter>
-      <Settings />
-    </HashRouter>
+    <Settings />
   </GlobalContext.Provider>
 );
