+++ assets/js/src/settings/pages/index.tsx
@@ -2,6 +2,5 @@ export { default as Basics } from './basics/basics';
 export { default as Advanced } from './advanced/advanced';
 export { default as KeyActivation } from './key_activation/key_activation';
 export { default as SendWith } from './send_with/send_with';
-export { default as OtherSendingMethods } from './send_with/other/other_sending_methods';
 export { default as SignupConfirmation } from './signup_confirmation/signup_confirmation';
 export { default as WooCommerce } from './woo_commerce/woo_commerce';
